+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas.geometry.Brep.cap_planar_holes`.
 * Added `compas_rhino.geometry.RhinoBrep.cap_planar_holes`.
 * Added `compas.geometry.angle_vectors_projected`.
+* Added `compas.geometry.Brep.from_curves`.
+* Added `compas_rhino.geometry.RhinoBrep.from_curves`.
 
 ### Changed